making sure all subscription utils are up to date with the new CustomTierLimit as well as slighly improving code and moving to different files
